dedpuplicate shared validation code
